+++ packages/host/app/services/render-service.ts
@@ -15,8 +15,9 @@ import type { RealmPaths } from '@cardstack/runtime-common';
 import {
   logger,
   baseRealm,
-  type CardStoreWithErrors,
+  loadDocument,
   type CodeRef,
+  type SingleCardDocument,
 } from '@cardstack/runtime-common';
 import type { Deferred } from '@cardstack/runtime-common/deferred';
 import { isCardError, CardError } from '@cardstack/runtime-common/error';
@@ -28,6 +29,7 @@ import config from '@cardstack/host/config/environment';
 import type {
   CardDef,
   Format,
+  CardStore,
   BoxComponent,
 } from 'https://cardstack.com/base/card-api';
 
@@ -45,6 +47,54 @@ const log = logger('renderer');
 
 const ELEMENT_NODE_TYPE = 1;
 const { environment } = config;
+
+export class CardStoreWithErrors implements CardStore {
+  #cards = new Map<string, CardDef>();
+  #fetch: typeof globalThis.fetch;
+  #inFlight: Promise<unknown>[] = [];
+  #docsInFlight: Map<string, Promise<SingleCardDocument | CardError>> =
+    new Map();
+
+  constructor(fetch: typeof globalThis.fetch) {
+    this.#fetch = fetch;
+  }
+
+  get(id: string): CardDef | undefined {
+    id = id.replace(/\.json$/, '');
+    return this.#cards.get(id);
+  }
+  set(id: string, instance: CardDef): void {
+    id = id.replace(/\.json$/, '');
+    this.#cards.set(id, instance);
+  }
+  setNonTracked(id: string, instance: CardDef) {
+    id = id.replace(/\.json$/, '');
+    return this.#cards.set(id, instance);
+  }
+  makeTracked(_id: string) {}
+
+  readonly errors = new Set<string>();
+
+  async loadDocument(url: string) {
+    let promise = this.#docsInFlight.get(url);
+    if (promise) {
+      return await promise;
+    }
+    try {
+      promise = loadDocument(this.#fetch, url);
+      this.#docsInFlight.set(url, promise);
+      return await promise;
+    } finally {
+      this.#docsInFlight.delete(url);
+    }
+  }
+  trackLoad(load: Promise<unknown>) {
+    this.#inFlight.push(load);
+  }
+  async loaded() {
+    await Promise.allSettled(this.#inFlight);
+  }
+}
 
 export interface RenderCardParams {
   card: CardDef;
