+++ packages/runtime-common/index.ts
@@ -160,7 +160,6 @@ export * from './index-query-engine';
 export * from './index-writer';
 export * from './index-structure';
 export * from './db';
-export * from './card-store-with-errors';
 export * from './tasks';
 export * from './worker';
 export * from './stream';
